refactor(components): reorganize GitHub component structure
- Move BatchRepoForm to src/components/github/forms/
- Move BatchResults to src/components/github/results/
- Move repo-analysis components to src/components/github/analysis/tabs/
- Update import paths accordingly
- Remove duplicate type definitions
