--- src/components/github/RepoResults.tsx
@@ -1,8 +1,0 @@
-import type { RepoData } from './repo-analysis/types';
-import ModularRepoResults from './repo-analysis/RepoResults';
-
-// Re-export component to maintain backward compatibility
-export default ModularRepoResults;
-
-// Export RepoData type for type compatibility
-export type { RepoData };
