+++ src/pages/GitHub.tsx
@@ -1,7 +1,7 @@
 import { Container, Box, Typography } from "@mui/material";
 import GitHubIcon from "@mui/icons-material/GitHub";
 import RepoAnalysisForm from "../components/github/RepoAnalysisForm";
-import RepoTabs from "../components/github/RepoTabs";
+import RepoTabs from "../components/github/analysis/tabs/RepoTabs.tsx";
 
 export const GitHub = (): React.ReactElement => {
 	return (
